pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "codeassist"
version = "0.1.0"
description = "LLM-Based Code Completion and Review Tool"
authors = [{name = "CodeAssist Team"}]
dependencies = [
    "click>=8.0.0",
    "rich>=13.0.0",
    "openai>=1.3.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.0.0",
]

[project.scripts]
codeassist = "codeassist.cli.commands:main"
codeassist-api = "codeassist.api.main:app"